refactor: setup the viewport to an extension
