+++ Sources/Clappr/Classes/Extension/AVFoundationPlayback+ViewPort.swift
@@ -1,0 +1,10 @@
+import AVFoundation
+
+extension AVFoundationPlayback {
+
+    @available(iOS 11.0, *)
+    @available(tvOS 11.0, *)
+    func setupMaxResolution(for size: CGSize) {
+        player?.currentItem?.preferredMaximumResolution = size
+    }
+}
