+++ Tests/Clappr_Tests/AVPlayerStub.swift
@@ -1,0 +1,14 @@
+import AVFoundation
+
+class AVPlayerStub: AVPlayer {
+
+    override var currentItem: AVPlayerItem? {
+        return _item
+    }
+
+    var _item = AVPlayerItemMock(url: URL(string: "https://clappr.io/highline.mp4")!)
+
+    func setStatus(to newStatus: AVPlayerItemStatus) {
+        _item._status = newStatus
+    }
+}
